Don't actually send metadata as JSON
This was thought of as a quick and easy way to do parsing for the server without writing on in Elixir, but it does not actually make much sense, since the server will have to do basic parsing anyways to understand which events are Game Start / Game End, in order to send new viewers the correct binary upon connection. If the server is already doing this, it just makes more sense to parse basic data from the Game Start event. This will help keep metadata synchronous, keep logic where it's expected, and make sending bad data more difficult.